+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/features/queue-management/components/CapacityEditorDialog.tsx
@@ -18,7 +18,7 @@
 
 
 import React from 'react';
-import { Loader2, Plus, X } from 'lucide-react';
+import { Loader2 } from 'lucide-react';
 import {
   Dialog,
   DialogContent,
@@ -28,16 +28,7 @@ import {
 } from '~/components/ui/dialog';
 import { Button } from '~/components/ui/button';
 import { Kbd } from '~/components/ui/kbd';
-import {
-  Field,
-  FieldLabel,
-  FieldDescription,
-  FieldControl,
-  FieldMessage,
-} from '~/components/ui/field';
-import { Badge } from '~/components/ui/badge';
-import { ToggleGroup, ToggleGroupItem } from '~/components/ui/toggle-group';
-import { Input } from '~/components/ui/input';
+import { Field, FieldLabel, FieldDescription, FieldControl } from '~/components/ui/field';
 import { ScrollArea } from '~/components/ui/scroll-area';
 import {
   Select,
@@ -48,7 +39,6 @@ import {
 } from '~/components/ui/select';
 import { useSchedulerStore } from '~/stores/schedulerStore';
 import type { SchedulerStore } from '~/stores/schedulerStore';
-import { cn } from '~/utils/cn';
 import { useKeyboardShortcuts, getModifierKey } from '~/hooks/useKeyboardShortcuts';
 import {
   DEFAULT_PARTITION_VALUE,
@@ -57,199 +47,13 @@ import {
   ensureCoreEntries,
   parseVectorDraft,
 } from '~/features/queue-management/utils/capacityEditor';
+import { computeRemainingHelper } from '../utils/capacityRemainingHelper';
+import { RemainingHelperDisplay } from './RemainingHelperDisplay';
+import { CapacityRowEditor } from './CapacityRowEditor';
 import type { CapacityResourceMode, CapacityRowDraft } from '~/stores/slices/capacityEditorSlice';
 import { SPECIAL_VALUES } from '~/types';
-import type { QueueCapacitiesByPartition } from '~/types';
-import { parseCapacityValue } from '~/utils/capacityUtils';
-import type { ValidationIssue } from '~/types';
 
 type VectorTarget = 'capacity' | 'maxCapacity';
-
-const SUPPORTED_ABSOLUTE_RESOURCES = ['memory', 'vcores'] as const;
-type SupportedAbsoluteResource = (typeof SUPPORTED_ABSOLUTE_RESOURCES)[number];
-
-type RemainingHelper =
-  | {
-      kind: 'percentage-legacy';
-      remaining: number;
-      target: number;
-      isOverOrUnder: boolean;
-    }
-  | {
-      kind: 'weight-legacy';
-      sum: number;
-    }
-  | {
-      kind: 'absolute-legacy';
-      resources: Array<{
-        resource: SupportedAbsoluteResource;
-        allocated: number;
-        remaining: number;
-        total: number;
-      }>;
-    };
-
-const formatNumber = (value: number): string => {
-  if (!Number.isFinite(value)) {
-    return '0';
-  }
-  const rounded = Math.round((value + Number.EPSILON) * 100) / 100;
-  const formatted = rounded.toString();
-  if (formatted.includes('.')) {
-    return formatted.replace(/\.0+$/, '').replace(/(\.\d*[1-9])0+$/, '$1');
-  }
-  return formatted;
-};
-
-const computeRemainingHelper = (
-  rows: CapacityRowDraft[],
-  parentCapacityValue: string,
-  isLegacyMode: boolean,
-  parentQueuePath: string | null,
-  selectedNodeLabel: string | null,
-  getQueuePartitionCapacities: (
-    path: string,
-    partition: string,
-  ) => QueueCapacitiesByPartition | null,
-): RemainingHelper | null => {
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const allSimple = rows.every((row) => row.mode === 'simple');
-  const allVector = rows.every((row) => row.mode === 'vector');
-
-  // Legacy Mode
-  if (isLegacyMode) {
-    if (allSimple) {
-      let determinedType: 'percentage' | 'weight' | null = null;
-      let currentTotal = 0;
-
-      for (const row of rows) {
-        const currentParsed = parseCapacityValue(row.capacityValue);
-        const baseParsed = parseCapacityValue(row.baseCapacityValue);
-
-        const candidateType =
-          currentParsed && (currentParsed.type === 'percentage' || currentParsed.type === 'weight')
-            ? currentParsed.type
-            : baseParsed && (baseParsed.type === 'percentage' || baseParsed.type === 'weight')
-              ? baseParsed.type
-              : null;
-
-        if (!candidateType) {
-          return null;
-        }
-
-        if (!determinedType) {
-          determinedType = candidateType;
-        } else if (determinedType !== candidateType) {
-          return null;
-        }
-
-        if (currentParsed?.type === determinedType) {
-          currentTotal += currentParsed.value;
-        }
-      }
-
-      if (!determinedType) {
-        return null;
-      }
-
-      if (determinedType === 'percentage') {
-        const target = 100;
-        const remaining = target - currentTotal;
-        return {
-          kind: 'percentage-legacy',
-          remaining,
-          target,
-          isOverOrUnder: remaining !== 0,
-        };
-      }
-
-      if (determinedType === 'weight') {
-        return {
-          kind: 'weight-legacy',
-          sum: currentTotal,
-        };
-      }
-    }
-
-    if (allVector) {
-      // Legacy mode with absolute resources: show remaining capacity
-      if (!parentQueuePath) {
-        return null;
-      }
-
-      const partitionName = selectedNodeLabel || '';
-      const partition = getQueuePartitionCapacities(parentQueuePath, partitionName);
-
-      if (!partition) {
-        return null;
-      }
-
-      // Determine if parent is root to choose the appropriate total
-      const isParentRoot = parentQueuePath === SPECIAL_VALUES.ROOT_QUEUE_NAME;
-
-      // Calculate allocated resources from all rows
-      const allocatedResources = new Map<string, number>();
-      rows.forEach((row) => {
-        row.vectorCapacity.forEach(({ key, value }) => {
-          if (key.trim().length === 0) {
-            return;
-          }
-          const numeric = Number.parseFloat(value);
-          if (!Number.isNaN(numeric)) {
-            const current = allocatedResources.get(key) ?? 0;
-            allocatedResources.set(key, current + numeric);
-          }
-        });
-      });
-
-      const resources: Array<{
-        resource: SupportedAbsoluteResource;
-        allocated: number;
-        remaining: number;
-        total: number;
-      }> = [];
-
-      SUPPORTED_ABSOLUTE_RESOURCES.forEach((resource) => {
-        const resourceKey = resource === 'vcores' ? 'vCores' : resource;
-
-        // For root, use effectiveMaxResource; for non-root, use configuredMinResource
-        const total = isParentRoot
-          ? (partition.effectiveMaxResource?.[resourceKey] ?? 0)
-          : (partition.configuredMinResource?.[resourceKey] ?? 0);
-
-        const allocated = allocatedResources.get(resource) ?? 0;
-        const remaining = total - allocated;
-
-        // Only include resources that have a total or allocation
-        if (total > 0 || allocated > 0) {
-          resources.push({
-            resource,
-            allocated,
-            remaining,
-            total,
-          });
-        }
-      });
-
-      if (resources.length === 0) {
-        return null;
-      }
-
-      return {
-        kind: 'absolute-legacy',
-        resources,
-      };
-    }
-
-    return null;
-  }
-
-  // Non-Legacy Mode: no strict validation rules, don't show helper
-  return null;
-};
 
 const useLegacyMode = (): boolean => {
   return useSchedulerStore(
@@ -300,14 +104,14 @@ export const CapacityEditorDialog: React.FC = () => {
     .map((queuePath) => drafts[queuePath])
     .filter((row): row is CapacityRowDraft => Boolean(row));
 
-  const remainingHelper = computeRemainingHelper(
+  const remainingHelper = computeRemainingHelper({
     rows,
     parentCapacityValue,
     isLegacyMode,
     parentQueuePath,
     selectedNodeLabel,
     getQueuePartitionCapacities,
-  );
+  });
 
   const hasBlockingIssues = validationIssues.some((issue) => issue.severity === 'error');
 
@@ -412,115 +216,6 @@ export const CapacityEditorDialog: React.FC = () => {
 
   const selectValue = selectedNodeLabel ?? DEFAULT_PARTITION_VALUE;
 
-  const renderVectorEntries = (
-    row: CapacityRowDraft,
-    target: VectorTarget,
-    issues: ValidationIssue[],
-  ) => {
-    const key = target === 'capacity' ? 'vectorCapacity' : 'vectorMaxCapacity';
-    const entries = row[key];
-    const headline = target === 'capacity' ? 'Capacity vector' : 'Maximum capacity vector';
-
-    return (
-      <Field>
-        <FieldLabel className="text-xs uppercase tracking-wide text-muted-foreground">
-          {headline}
-        </FieldLabel>
-        {issues.length === 0 && (
-          <>
-            <FieldDescription className="text-xs text-muted-foreground">
-              {isLegacyMode
-                ? 'Legacy mode: enter numeric values per resource.'
-                : 'Use numeric values for counts, append w for weights and % for percentages.'}
-            </FieldDescription>
-            <FieldDescription className="text-[11px] text-muted-foreground">
-              Base:{' '}
-              {target === 'capacity'
-                ? row.baseCapacityValue || '—'
-                : row.baseMaxCapacityValue || '—'}
-            </FieldDescription>
-          </>
-        )}
-        <div className="mt-2 space-y-2">
-          {entries.map((entry, index) => {
-            const isCoreResource = entry.key === 'memory' || entry.key === 'vcores' || index < 2;
-            return (
-              <div key={entry.id} className="flex items-center gap-2">
-                <Input
-                  value={entry.key}
-                  onChange={(event) =>
-                    handleVectorEntryChange(
-                      row.queuePath,
-                      target,
-                      entry.id,
-                      'key',
-                      event.target.value,
-                    )
-                  }
-                  placeholder="resource"
-                  className="h-8 w-32 text-sm"
-                />
-                <Input
-                  value={entry.value}
-                  onChange={(event) =>
-                    handleVectorEntryChange(
-                      row.queuePath,
-                      target,
-                      entry.id,
-                      'value',
-                      event.target.value,
-                    )
-                  }
-                  placeholder={isLegacyMode ? '0' : '0 | 50% | 10w'}
-                  className="h-8 text-sm"
-                />
-                {!isCoreResource && (
-                  <Button
-                    type="button"
-                    size="icon"
-                    variant="ghost"
-                    className="h-8 w-8 text-muted-foreground"
-                    onClick={() => handleRemoveVectorEntry(row.queuePath, target, entry.id)}
-                  >
-                    <X className="h-4 w-4" />
-                  </Button>
-                )}
-              </div>
-            );
-          })}
-        </div>
-        <div className="mt-2">
-          <Button
-            type="button"
-            variant="ghost"
-            size="sm"
-            className="text-xs"
-            onClick={() => handleAddVectorEntry(row.queuePath, target)}
-          >
-            <Plus className="mr-2 h-3.5 w-3.5" />
-            Add resource
-          </Button>
-        </div>
-        {issues.length > 0 && (
-          <div className="mt-2 space-y-1">
-            {issues.map((issue) => (
-              <FieldMessage
-                key={`${issue.rule}-${issue.field}`}
-                className={
-                  issue.severity === 'error'
-                    ? 'text-[11px] text-destructive'
-                    : 'text-[11px] text-amber-600'
-                }
-              >
-                {issue.message}
-              </FieldMessage>
-            ))}
-          </div>
-        )}
-      </Field>
-    );
-  };
-
   if (!isOpen) {
     return null;
   }
@@ -567,40 +262,7 @@ export const CapacityEditorDialog: React.FC = () => {
             </Field>
           </div>
 
-          {remainingHelper && (
-            <div
-              className={cn(
-                'mt-3 rounded-md border border-dashed px-3 py-2 text-xs text-left space-y-1',
-                (remainingHelper.kind === 'percentage-legacy' && remainingHelper.isOverOrUnder) ||
-                  (remainingHelper.kind === 'absolute-legacy' &&
-                    remainingHelper.resources.some((r) => r.remaining < 0))
-                  ? 'bg-amber-50/60 border-amber-500/60 text-amber-900'
-                  : 'bg-muted/40 text-muted-foreground',
-              )}
-            >
-              {remainingHelper.kind === 'percentage-legacy' && (
-                <p>
-                  {remainingHelper.remaining >= 0
-                    ? `${formatNumber(remainingHelper.remaining)}% capacity remaining`
-                    : `${formatNumber(Math.abs(remainingHelper.remaining))}% over target`}{' '}
-                  (target {formatNumber(remainingHelper.target)}%)
-                </p>
-              )}
-              {remainingHelper.kind === 'weight-legacy' && (
-                <p>Sum of weights: {formatNumber(remainingHelper.sum)}</p>
-              )}
-              {remainingHelper.kind === 'absolute-legacy' &&
-                remainingHelper.resources.map((resource) => (
-                  <p key={resource.resource}>
-                    {resource.resource}:{' '}
-                    {resource.remaining >= 0
-                      ? `${formatNumber(resource.remaining)} remaining`
-                      : `${formatNumber(Math.abs(resource.remaining))} over target`}{' '}
-                    (total {formatNumber(resource.total)})
-                  </p>
-                ))}
-            </div>
-          )}
+          {remainingHelper && <RemainingHelperDisplay helper={remainingHelper} />}
         </DialogHeader>
 
         <ScrollArea className="max-h-[60vh] pr-3">
@@ -620,158 +282,32 @@ export const CapacityEditorDialog: React.FC = () => {
                 (issue) => issue.queuePath === row.queuePath && issue.field === maxFieldName,
               );
 
-              const hasRowError =
-                capacityIssuesForRow.some((issue) => issue.severity === 'error') ||
-                maxIssuesForRow.some((issue) => issue.severity === 'error');
-
               return (
-                <div
+                <CapacityRowEditor
                   key={row.queuePath}
-                  className={cn(
-                    'rounded-md border p-4 transition',
-                    row.isOrigin && 'border-primary/70 bg-primary/5',
-                    row.hasStagedChange && !row.isOrigin && 'border-amber-500/60 bg-amber-50/60',
-                    row.isNew && !row.isOrigin && 'border-dashed',
-                    hasRowError && 'border-destructive/70 bg-destructive/5',
-                  )}
-                >
-                  <div className="flex flex-wrap items-start justify-between gap-3">
-                    <div className="space-y-1">
-                      <div className="flex flex-wrap items-center gap-2">
-                        <span className="text-sm font-medium text-foreground">{row.queueName}</span>
-                        {row.isOrigin && (
-                          <Badge variant="secondary" className="h-4 px-1 text-[10px]">
-                            Active queue
-                          </Badge>
-                        )}
-                        {row.hasStagedChange && !row.isOrigin && (
-                          <Badge variant="outline" className="h-4 px-1 text-[10px]">
-                            Staged
-                          </Badge>
-                        )}
-                        {row.isNew && !row.isOrigin && (
-                          <Badge variant="secondary" className="h-4 px-1 text-[10px]">
-                            New
-                          </Badge>
-                        )}
-                      </div>
-                      <p className="break-all text-[11px] text-muted-foreground">{row.queuePath}</p>
-                    </div>
-
-                    <ToggleGroup
-                      type="single"
-                      value={row.mode}
-                      onValueChange={(value) =>
-                        handleModeChange(row.queuePath, value as CapacityResourceMode)
-                      }
-                      className="shrink-0"
-                      variant="outline"
-                    >
-                      <ToggleGroupItem value="simple" className="text-xs px-3 py-1.5">
-                        Simple value
-                      </ToggleGroupItem>
-                      <ToggleGroupItem value="vector" className="text-xs px-3 py-1.5">
-                        Resource vector
-                      </ToggleGroupItem>
-                    </ToggleGroup>
-                  </div>
-
-                  {row.mode === 'simple' ? (
-                    <div className="mt-4 grid gap-3 md:grid-cols-2">
-                      <Field>
-                        <FieldLabel className="text-xs uppercase tracking-wide text-muted-foreground">
-                          Capacity
-                        </FieldLabel>
-                        <FieldControl>
-                          <Input
-                            value={row.capacityValue}
-                            onChange={(event) => {
-                              updateCapacityDraft(row.queuePath, (draft) => {
-                                draft.capacityValue = event.target.value;
-                              });
-                            }}
-                            placeholder="e.g. 50, 10w"
-                            className="h-8 text-sm"
-                          />
-                        </FieldControl>
-                        {capacityIssuesForRow.length === 0 && (
-                          <>
-                            <FieldDescription className="text-[11px] text-muted-foreground">
-                              {`Use numbers for percentages (e.g. 50) or append w for weights (e.g. 10w).`}
-                            </FieldDescription>
-                            <FieldDescription className="text-[11px] text-muted-foreground">
-                              Base: {row.baseCapacityValue || '—'}
-                            </FieldDescription>
-                          </>
-                        )}
-                        {capacityIssuesForRow.length > 0 && (
-                          <div className="mt-2 space-y-1">
-                            {capacityIssuesForRow.map((issue) => (
-                              <FieldMessage
-                                key={`${issue.rule}-${issue.field}`}
-                                className={
-                                  issue.severity === 'error'
-                                    ? 'text-[11px] text-destructive'
-                                    : 'text-[11px] text-amber-600'
-                                }
-                              >
-                                {issue.message}
-                              </FieldMessage>
-                            ))}
-                          </div>
-                        )}
-                      </Field>
-                      <Field>
-                        <FieldLabel className="text-xs uppercase tracking-wide text-muted-foreground">
-                          Maximum capacity
-                        </FieldLabel>
-                        <FieldControl>
-                          <Input
-                            value={row.maxCapacityValue}
-                            onChange={(event) => {
-                              updateCapacityDraft(row.queuePath, (draft) => {
-                                draft.maxCapacityValue = event.target.value;
-                              });
-                            }}
-                            placeholder="e.g. 100, 20w"
-                            className="h-8 text-sm"
-                          />
-                        </FieldControl>
-                        {maxIssuesForRow.length === 0 && (
-                          <>
-                            <FieldDescription className="text-[11px] text-muted-foreground">
-                              {`Maximum value the queue can reach.`}
-                            </FieldDescription>
-                            <FieldDescription className="text-[11px] text-muted-foreground">
-                              Base: {row.baseMaxCapacityValue || '—'}
-                            </FieldDescription>
-                          </>
-                        )}
-                        {maxIssuesForRow.length > 0 && (
-                          <div className="mt-2 space-y-1">
-                            {maxIssuesForRow.map((issue) => (
-                              <FieldMessage
-                                key={`${issue.rule}-${issue.field}`}
-                                className={
-                                  issue.severity === 'error'
-                                    ? 'text-[11px] text-destructive'
-                                    : 'text-[11px] text-amber-600'
-                                }
-                              >
-                                {issue.message}
-                              </FieldMessage>
-                            ))}
-                          </div>
-                        )}
-                      </Field>
-                    </div>
-                  ) : (
-                    <div className="mt-4 space-y-4">
-                      {renderVectorEntries(row, 'capacity', capacityIssuesForRow)}
-                      {renderVectorEntries(row, 'maxCapacity', maxIssuesForRow)}
-                    </div>
-                  )}
-                </div>
+                  row={row}
+                  capacityIssues={capacityIssuesForRow}
+                  maxCapacityIssues={maxIssuesForRow}
+                  isLegacyMode={isLegacyMode}
+                  onModeChange={(mode) => handleModeChange(row.queuePath, mode)}
+                  onCapacityChange={(value) => {
+                    updateCapacityDraft(row.queuePath, (draft) => {
+                      draft.capacityValue = value;
+                    });
+                  }}
+                  onMaxCapacityChange={(value) => {
+                    updateCapacityDraft(row.queuePath, (draft) => {
+                      draft.maxCapacityValue = value;
+                    });
+                  }}
+                  onVectorEntryChange={(target, entryId, field, value) =>
+                    handleVectorEntryChange(row.queuePath, target, entryId, field, value)
+                  }
+                  onAddVectorEntry={(target) => handleAddVectorEntry(row.queuePath, target)}
+                  onRemoveVectorEntry={(target, entryId) =>
+                    handleRemoveVectorEntry(row.queuePath, target, entryId)
+                  }
+                />
               );
             })}
           </div>
