+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/features/validation/__tests__/crossQueue.test.ts
@@ -18,11 +18,7 @@
 
 
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import {
-  validatePropertyChange,
-  validateAllStagedChanges,
-  selectivelyValidateStagedChanges,
-} from '~/features/validation/crossQueue';
+import { validatePropertyChange, validateStagedChanges } from '~/features/validation/crossQueue';
 import type { SchedulerInfo, StagedChange, ValidationIssue } from '~/types';
 
 // Mock dependencies
@@ -639,9 +635,9 @@ describe('crossQueue validation', () => {
     });
   });
 
-  describe('validateAllStagedChanges', () => {
+  describe('validateStagedChanges', () => {
     it('should return empty map when schedulerData is null', () => {
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges: [],
         schedulerData: null,
         configData: new Map(),
@@ -651,7 +647,7 @@ describe('crossQueue validation', () => {
     });
 
     it('should return empty map when staged changes is empty', () => {
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges: [],
         schedulerData: createMockSchedulerData(),
         configData: new Map(),
@@ -688,7 +684,7 @@ describe('crossQueue validation', () => {
         ],
       });
 
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges,
         schedulerData,
         configData: new Map(),
@@ -718,7 +714,7 @@ describe('crossQueue validation', () => {
         issues: [],
       });
 
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges,
         schedulerData,
         configData: new Map(),
@@ -742,7 +738,7 @@ describe('crossQueue validation', () => {
         },
       ];
 
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges,
         schedulerData,
         configData: new Map(),
@@ -776,7 +772,7 @@ describe('crossQueue validation', () => {
         },
       ];
 
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges,
         schedulerData,
         configData: new Map(),
@@ -810,7 +806,7 @@ describe('crossQueue validation', () => {
 
       vi.mocked(getAffectedQueuesForValidation).mockReturnValue(['root.default']);
 
-      validateAllStagedChanges({
+      validateStagedChanges({
         stagedChanges,
         schedulerData,
         configData: new Map(),
@@ -843,7 +839,7 @@ describe('crossQueue validation', () => {
         issues: [],
       });
 
-      const result = validateAllStagedChanges({
+      const result = validateStagedChanges({
         stagedChanges,
         schedulerData,
         configData: new Map(),
@@ -853,9 +849,9 @@ describe('crossQueue validation', () => {
     });
   });
 
-  describe('selectivelyValidateStagedChanges', () => {
+  describe('validateStagedChanges', () => {
     it('should return empty map when schedulerData is null', () => {
-      const result = selectivelyValidateStagedChanges({
+      const result = validateStagedChanges({
         affectedQueuePaths: new Set(),
         affectedProperties: new Set(),
         stagedChanges: [],
@@ -867,7 +863,7 @@ describe('crossQueue validation', () => {
     });
 
     it('should return empty map when staged changes is empty', () => {
-      const result = selectivelyValidateStagedChanges({
+      const result = validateStagedChanges({
         affectedQueuePaths: new Set(),
         affectedProperties: new Set(),
         stagedChanges: [],
@@ -898,7 +894,7 @@ describe('crossQueue validation', () => {
         issues: [],
       });
 
-      const result = selectivelyValidateStagedChanges({
+      const result = validateStagedChanges({
         affectedQueuePaths: new Set(['root.default']),
         affectedProperties: new Set(),
         stagedChanges,
@@ -930,7 +926,7 @@ describe('crossQueue validation', () => {
         issues: [],
       });
 
-      const result = selectivelyValidateStagedChanges({
+      const result = validateStagedChanges({
         affectedQueuePaths: new Set(),
         affectedProperties: new Set(['capacity']),
         stagedChanges,
@@ -966,7 +962,7 @@ describe('crossQueue validation', () => {
         },
       ];
 
-      const result = selectivelyValidateStagedChanges({
+      const result = validateStagedChanges({
         affectedQueuePaths: new Set(['root.default']),
         affectedProperties: new Set(),
         stagedChanges,
@@ -1003,7 +999,7 @@ describe('crossQueue validation', () => {
 
       vi.mocked(getAffectedQueuesForValidation).mockReturnValue(['root.default']);
 
-      selectivelyValidateStagedChanges({
+      validateStagedChanges({
         affectedQueuePaths: new Set(['root.default']),
         affectedProperties: new Set(),
         stagedChanges,
@@ -1035,7 +1031,7 @@ describe('crossQueue validation', () => {
         issues: [],
       });
 
-      const result = selectivelyValidateStagedChanges({
+      const result = validateStagedChanges({
         affectedQueuePaths: new Set(['root.newqueue']),
         affectedProperties: new Set(),
         stagedChanges,
