+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/features/validation/utils/affectedQueues.ts
@@ -18,7 +18,7 @@
 
 
 import type { SchedulerInfo, StagedChange } from '~/types';
-import { findQueueByPath } from '~/utils/queueTreeUtils';
+import { findQueueByPath } from '~/utils/treeUtils';
 import { getParentQueuePath } from '~/utils/propertyUtils';
 import { isTemplateQueuePath } from '~/utils/templateUtils';
