+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/types/guards.ts
@@ -28,16 +28,35 @@ export function isParentQueue(queue: QueueInfo): boolean {
   return queue.queueType === QUEUE_TYPES.PARENT;
 }
 
-export function isValidQueueName(name: string): boolean {
-  if (!name || name.trim() !== name) {
-    return false;
+/**
+ * Returns an error message if the queue name is invalid, or null if valid.
+ * This is the single source of truth for queue name validation rules.
+ */
+export function getQueueNameValidationError(name: string): string | null {
+  if (!name || name.trim() === '') {
+    return 'Queue name cannot be empty';
   }
 
-  // Queue names cannot contain dots
+  if (name.trim() !== name) {
+    return 'Queue name cannot have leading or trailing whitespace';
+  }
+
+  // Queue names cannot contain dots (they are used as path separators)
   if (name.includes('.')) {
-    return false;
+    return 'Queue names cannot contain dots (.)';
   }
 
   // Must match alphanumeric, hyphen, underscore pattern
-  return /^[a-zA-Z0-9_-]+$/.test(name);
+  if (!/^[a-zA-Z0-9_-]+$/.test(name)) {
+    return 'Queue names should only contain letters, numbers, hyphens, and underscores';
+  }
+
+  return null;
+}
+
+/**
+ * Returns true if the queue name is valid.
+ */
+export function isValidQueueName(name: string): boolean {
+  return getQueueNameValidationError(name) === null;
 }
