+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/utils/capacityUtils.ts
@@ -20,6 +20,62 @@
 import type { CapacityType, ParsedCapacity } from '~/types';
 
 /**
+ * Entry in a resource vector (e.g., memory=1024, vcores=2)
+ */
+export interface ResourceVectorEntry {
+  resource: string;
+  value: string;
+}
+
+/**
+ * Check if a value looks like a resource vector (starts with [ and ends with ])
+ * @param value The value to check
+ * @returns True if the value appears to be a vector format
+ */
+export function isVectorCapacity(value: string | undefined | null): boolean {
+  if (!value) {
+    return false;
+  }
+  const trimmed = value.trim();
+  return trimmed.startsWith('[') && trimmed.endsWith(']');
+}
+
+/**
+ * Parse a resource vector string into an array of resource entries
+ * @param value Resource vector string (e.g., "[memory=1024,vcores=2]")
+ * @returns Array of resource entries, or empty array if invalid
+ */
+export function parseResourceVector(value: string): ResourceVectorEntry[] {
+  const trimmed = value.trim();
+  if (!isVectorCapacity(trimmed)) {
+    return [];
+  }
+
+  const inner = trimmed.slice(1, -1).trim();
+  if (!inner) {
+    return [];
+  }
+
+  return inner
+    .split(',')
+    .map((pair) => {
+      const [resource, val] = pair.split('=');
+      const resourceName = resource?.trim();
+      const resourceValue = val?.trim();
+
+      if (!resourceName || !resourceValue) {
+        return null;
+      }
+
+      return {
+        resource: resourceName,
+        value: resourceValue,
+      };
+    })
+    .filter((entry): entry is ResourceVectorEntry => entry !== null);
+}
+
+/**
  * Parse a capacity value string into a structured format
  * Supports: percentages (50%), weights (5w), absolute resources ([memory=1024,vcores=2])
  * @param value Capacity value string
