+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/utils/formatUtils.ts
@@ -17,8 +17,6 @@
  */
 
 
-import type { ResourceInfo } from '~/types/resource';
-import type { QueueInfo } from '~/types/queue';
 import { SPECIAL_VALUES } from '~/types/constants/special-values';
 
 /**
