+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-capacity-scheduler-ui/src/main/webapp/src/utils/treeUtils.test.ts
@@ -1,27 +1,8 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
 import { describe, it, expect, vi } from 'vitest';
-import { queueTreeUtils } from './queueTreeUtils';
+import { flattenQueueTree, traverseQueueTree, findQueueByPath } from './treeUtils';
 import type { QueueInfo } from '~/types';
 
-describe('queueTreeUtils', () => {
+describe('treeUtils', () => {
   // Helper to create test queue tree
   const createTestQueueTree = (): QueueInfo => ({
     queueName: 'root',
@@ -110,7 +91,7 @@ describe('queueTreeUtils', () => {
   describe('flattenQueueTree', () => {
     it('should flatten a queue tree into an array', () => {
       const root = createTestQueueTree();
-      const flattened = queueTreeUtils.flattenQueueTree(root);
+      const flattened = flattenQueueTree(root);
 
       expect(flattened).toHaveLength(5);
       expect(flattened.map((q) => q.queuePath)).toEqual([
@@ -139,7 +120,7 @@ describe('queueTreeUtils', () => {
         state: 'RUNNING',
       };
 
-      const flattened = queueTreeUtils.flattenQueueTree(singleQueue);
+      const flattened = flattenQueueTree(singleQueue);
       expect(flattened).toHaveLength(1);
       expect(flattened[0]).toBe(singleQueue);
     });
@@ -164,7 +145,7 @@ describe('queueTreeUtils', () => {
         },
       };
 
-      const flattened = queueTreeUtils.flattenQueueTree(queueWithEmptyQueues);
+      const flattened = flattenQueueTree(queueWithEmptyQueues);
       expect(flattened).toHaveLength(1);
     });
   });
@@ -174,7 +155,7 @@ describe('queueTreeUtils', () => {
       const root = createTestQueueTree();
       const callback = vi.fn();
 
-      queueTreeUtils.traverseQueueTree(root, callback);
+      traverseQueueTree(root, callback);
 
       expect(callback).toHaveBeenCalledTimes(5);
 
@@ -200,7 +181,7 @@ describe('queueTreeUtils', () => {
       const root = createTestQueueTree();
       const parents: Array<QueueInfo | undefined> = [];
 
-      queueTreeUtils.traverseQueueTree(root, (_1, _2, parent) => {
+      traverseQueueTree(root, (_1, _2, parent) => {
         parents.push(parent);
       });
 
@@ -216,7 +197,7 @@ describe('queueTreeUtils', () => {
     it('should find queue by path', () => {
       const root = createTestQueueTree();
 
-      const found = queueTreeUtils.findQueueByPath(root, 'root.production.team1');
+      const found = findQueueByPath(root, 'root.production.team1');
       expect(found).not.toBeNull();
       expect(found?.queueName).toBe('team1');
       expect(found?.queuePath).toBe('root.production.team1');
@@ -225,21 +206,21 @@ describe('queueTreeUtils', () => {
     it('should find root queue', () => {
       const root = createTestQueueTree();
 
-      const found = queueTreeUtils.findQueueByPath(root, 'root');
+      const found = findQueueByPath(root, 'root');
       expect(found).toBe(root);
     });
 
     it('should return null for non-existent path', () => {
       const root = createTestQueueTree();
 
-      const found = queueTreeUtils.findQueueByPath(root, 'root.nonexistent');
+      const found = findQueueByPath(root, 'root.nonexistent');
       expect(found).toBeNull();
     });
 
     it('should return null for partial path', () => {
       const root = createTestQueueTree();
 
-      const found = queueTreeUtils.findQueueByPath(root, 'production');
+      const found = findQueueByPath(root, 'production');
       expect(found).toBeNull();
     });
 
@@ -260,10 +241,10 @@ describe('queueTreeUtils', () => {
         state: 'RUNNING',
       };
 
-      const found = queueTreeUtils.findQueueByPath(singleQueue, 'root');
+      const found = findQueueByPath(singleQueue, 'root');
       expect(found).toBe(singleQueue);
 
-      const notFound = queueTreeUtils.findQueueByPath(singleQueue, 'root.child');
+      const notFound = findQueueByPath(singleQueue, 'root.child');
       expect(notFound).toBeNull();
     });
   });
